src/tests/requirements.txt
alembic==1.7.3
arrow==1.2.0
attrs==21.2.0
bcrypt==3.2.0
bidict==0.21.3
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
croniter==1.0.15
dynaconf==3.1.7
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SocketIO==5.1.1
Flask-SSE==1.0.0
greenlet==1.1.1
idna==3.2
iexfinance==0.5.0
importlib-resources==5.2.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
Mako==1.1.5
MarkupSafe==2.0.1
multitasking==0.0.9
mysqlclient==2.0.3
nulltype==2.3.1
numpy==1.21.2
packaging==21.3
pandas==1.3.3
plaid-python==8.3.0
pluggy==1.0.0
py==1.11.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
python-engineio==4.2.1
python-socketio==5.4.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
rq==1.10.0
rq-dashboard==0.6.1
rq-scheduler==0.11.0
six==1.16.0
SQLAlchemy==1.4.25
sseclient-py==1.7.2
toml==0.10.2
urllib3==1.26.7
Werkzeug==2.0.1
yfinance==0.1.63
zipp==3.5.0
